+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change Log
 
 ## [0.2.0](https://github.com/dldevinc/django-awesomplete/tree/v0.2.0) - 2020-10-20
+- The *minchars* and *maxitems* arguments of AwesompleteWidget 
+are marked as obsolete and will be renamed to *min_chars* and *max_items*
+respectively.
 - Update dev environment